requirements.txt
openmeteo-requests
requests
requests-cache
retry-requests
numpy
pandas
beautifulsoup4
Pillow
tqdm
